use adw theme

+++ scripts/webview.ts
@@ -1,18 +1,27 @@
 import { SizeHint, Webview } from "jsr:@webview/webview@0.9.0";
 import { removeAllVirtualSinks } from "../lib.ts";
+import { AdwApp } from "jsr:@sigmasd/adw-app";
 
-const worker = new Worker(import.meta.resolve("./start.ts"), {
-  type: "module",
-});
-const webview = new Webview();
-webview.title = "Record Apps";
-webview.size = { width: 1000, height: 600, hint: SizeHint.NONE };
+if (import.meta.main) {
+  const worker = new Worker(import.meta.resolve("./start.ts"), {
+    type: "module",
+  });
+  const port = await new Promise<number>((resolve) => {
+    worker.onmessage = (e) => {
+      const { port } = e.data;
+      resolve(port);
+    };
+  });
 
-worker.onmessage = async (e) => {
-  const { port } = e.data;
-  webview.navigate(`http://localhost:${port}`);
-  webview.run();
+  const app = new AdwApp({ id: "io.github.sigmasd.recordapps" });
+  app.run((window) => {
+    const webview = new Webview(false, undefined, window);
+    webview.title = "Record Apps";
+    webview.size = { width: 1000, height: 600, hint: SizeHint.NONE };
+
+    webview.navigate(`http://localhost:${port}`);
+  });
   worker.terminate();
   await removeAllVirtualSinks();
   Deno.exit(0);
-};
+}
